+++ dhis-2/dhis-web/dhis-web-light/src/main/resources/META-INF/dhis/beans.xml
@@ -24,8 +24,7 @@ http://www.springframework.org/schema/beans http://www.springframework.org/schem
 		<property name="minMaxOutlierAnalysisService"
 			ref="org.hisp.dhis.dataanalysis.MinMaxOutlierAnalysisService" />
 		<property name="systemSettingManager" ref="org.hisp.dhis.setting.SystemSettingManager" />
-		<property name="validationRuleService"
-			ref="org.hisp.dhis.validation.ValidationRuleService" />
+		<property name="validationService" ref="org.hisp.dhis.validation.ValidationService" />
 		<property name="expressionService" ref="org.hisp.dhis.expression.ExpressionService" />
 	</bean>
